Re-work how trait implementations are tagged as 'effectively_blanket'
The goal is to recognise trait implementations like:

```
impl<'a> SomeTrait for SomeType<'a> {}
```

where the trait itself has no generic parameters and there isn't really
anything notable about the `impl<>` parameters since they are all just
passed to the `for` type.

In this case we use 'simple' formatting when listing trait
implementations so we would just list `- SomeTrait`

TODO: The current implementation has huge amount of nesting and right
drift which would be good to clean up!